# src/apps/crawler-server/openapi.yaml
openapi: 3.0.0
info:
  title: "Web crawler"
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /records:
    get:
      summary: Returns a list of user identifications.
      parameters:
        - in: query
          name: sortBy
          description: Name of field to sort by.
          schema:
            type: string
            # ENUM?
        # - in: query
        #   name: limit
        #   description: Maximum number of items in result (max 100; default = 100).
        #   schema:
        #     type: integer
        # - in: query
        #   name: offset
        #   description: Offset from the beginning of the data (sorted based on sortBy parameter; default = 0).
        #   schema:
        #     type: integer
        - in: query
          name: filterByURL
          description: URL to filter by (exact match required).
          schema:
            type: string
        - in: query
          name: filterByLabel
          description: Label to filter by (exact match required).
          schema:
            type: string
        - in: query
          name: filterByTags
          description: Tags to filter by (every record in the result will have ALL of the given tags).
          schema:
            type: array
            items: { type: string }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  # totalCount:
                  #   type: integer
                  #   description: Total count of the website records.
                  #   example: 120
                  data:
                    type: array
                    description: Requested array of website record ids.
                    items: { $ref: "#/components/schemas/websiteRecord" }
  /records/create:
    post:
      summary: Create a new website record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/websiteRecordChange"
      responses:
        "201":
          description: "Website record created"
          content:
            application/json:
              schema:
                type: object
                properties:
                  recordId:
                    type: string
                    description: ID of the new website record.
...
components:
  schemas:
    websiteRecord:
      type: object
      required:
        [
          id,
          url,
          boundaryRegex,
          periodicityInSeconds,
          label,
          isActive,
          tags,
          lastExecutionId,
        ]
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the website record.
        url:
          type: string
          description: URL of the webiste to crawl.
        boundaryRegex:
          type: string
          description: Regular expression describing the website domain's to crawl.
        periodicityInSeconds:
          type: integer
          description: Crawl periodicity in seconds (value 0 implies no automatic repetition).
        label:
          type: string
          description: Human readable label of the website record (non-empty string).
        isActive:
          type: boolean
          description: Boolean indication whether the website record is active.
        tags:
          type: array
          description: Array of tags of the website record.
          items:
            type: string
            description: Non-empty string
            example: "books"
        lastExecutionId:
          type: string
          description: Id of the last execution of the crawl described in the website record.
    # schema for representing creation or update of a website record
    websiteRecordChange:
      type: object
      properties:
        url:
          type: string
          description: URL of the webiste to crawl.
        boundaryRegex:
          type: string
          description: Regular expression describing the website domain's to crawl.
        periodicityInSeconds:
          type: integer
          description: Crawl periodicity in seconds (setting this to 0 implies no automatic repetition).
        label:
          type: string
          description: Human readable label of the website record (non-empty string).
        isActive:
          type: boolean
          description: Boolean indication whether the website record is active.
        tags:
          type: array
          description: Array of tags of the website record.
          items:
            type: string
            description: Non-empty string
            example: "books"
        lastExecutionURL:
          type: string
          description: URL of the last execution of the crawl described in the website record.
...